Function sunrise_libtimezone::utils::increment_overflow

#[must_use]pub fn increment_overflow<T: Num + CheckedAdd + Copy>(ip: &mut T, j: T) -> bool

Increment the given ip with j if it doesn't overflow.

If the operation overflow, this return true otherwise false.